Allow cycling between windows of an active app
Jumping to apps with hotkeys is still similar. In the previous version,
nothing happened when the hotkey of a certain app is triggered when the app
was already active.

This update allows me to cycle between all windows of an active app when
the same hotkey is used.

The overall logic is as follows:

- If the app is not running, launch it
- If the app is running but not active, focus on it
- If the app is running and active but has a single window, return
- If the app is running, active and has multiple windows, jump to another
  window